Batch applied changes up to r118113 from Omni's svn.
The highlights of the changes follow.

Configurations:
	- Switched to dwarf-with-dsym for valgrind
	- Added a series of Library configurations

OmniBase:
	- Fixed clang-sa issues
	- Split up out the C NSError helper functions into their own file for inclusion in Quick Look / Spotlight bundles.
	- Added option to include backtrace where NSErrors are generated (at least with the OB calls)
	- Added/improved runtime checking for common method signature errors:
		- Subclass signature differs from superclass or attached protocol (possible 64-bit issue)
		- Implementing deprecated methods, including support for defining them via protocols including 'Deprecated' in their name.
	- Whitelist some more Cocoa methods.
	- Added support for running code coverage from unit tests, though it uses Omni-internal scripts right now.
	- Added macros for the clang attributes {ns,cf}_returns_retained
	- Added OBBuildByCompilerVersion global CFStringRef
	- Added OB_UNUSED_VALUE macro to help hint to clang that a dead store is OK.
	- Updated assertion macros to cooperate with clang-sa, using the analyzer_noreturn attribute.
	- Added OBASSERT_NULL and OBASSERT_NOTNULL macros to help inform clang-sa of guarantees we make.

OmniFoundation:
	- Fixed clang-sa issues
	- Removed pre-10.5 compatability code
	- Fixed some 64-bit issues
	- Added CF callback helpers for NS{U,}Integer types
	- Added SHA-256 signature support and cleaned up NSData signature generation.
	- Removed some whole classes and various other methods we no longer use (OFBitField, OFOid)
	- A bit of progress on GC support, though more work remains
	- Various other cleanup & modernization in surrounding code
	- Added CDSA helpers for dealing with certificates, signatures and the like
	- Various improvements/cleanup in OFFilterProcess. Switched NSData-filtering methods to use OFFilterProcess
	- Factored OFXMLParser out of OFXMLDocument
	- Added OFXMLReader for a xmlReader-based interface. This avoids reading the entire XML document into DOM-form.
	- Added OFXMLInternedStringTable
	- Added OFXMLMaker/OFXMLTextWriter/OFXMLCFXMLTreeSink
	- Added OFXMLQName for namespace-awareness; started process of updating APIs to use QNames.
	- OFXMLBuffer now stores UTF-8 for simpler libxml interfacing and adds some quoting support.
	- Added OFXMLSignature with a ton of support for ASN.1, X.509 and various other numbers and letters.
	- Fixed bugs and improved performance in XML date reading/writing; added several tests.
	- Added some quarantine/volume extensions to NSFileManager
	- Extend NSUndoManager debug logging support to show disable/enable calls
	- Added/extended many tests.
	- Some updates to OFRandom, but these will be deprecated soon.

OmniSoftwareUpdate:
	- Fixed clang-sa and 64-bit issues
	- Added support for signed software update feeds
	- Added support for signature validation of downloads
	- Improved and simplified user interface for installing updates
	- Improved the installation process to deal with various error cases, quarantine
	- Added support for the user to mark an update as ignored.
	- Various other fixes and cleanup
	- Send the list of GL extensions in the server query in a bitfield, making it vastly smaller.

OmniDataObjects:
	- Fixed clang-sa issues
	- Make -[ODOAttribute isPrimaryKey] public
	- Add configurable support for synchronous writes, temp_store pragmas and vacuum on disconnect
	- Added ODOEditingContextWillSaveNotification
	- Renamed ODOEditingContext{Will,Did}Reset to have a Notification suffix
	- Added -awakeFromUnarchive, though it isn't called by ODO.
	- Added (currently disabled) support for lazy @dynamic method registration.
	- Fixed bug where -setDefaultAttributeValues wasn't issuing KVO.
	- Fixed Ruby->ObjC model generator @class issues and clang-sa issues.

OmniAppKit:
	- Fixed clang-sa issues
	- Removed pre-10.5 compatibility code
	- Make string globals 'NSString * const'.
	- Cleanup display of the user-visible version number in the About panel, dropping display of the svn revision.
	- Fixed bugs in the sheet-queuing support
	- Fixed various other bugs, added various other features

OmniInspector:
	- Fixed clang-sa and 64-bit issues
	- A few other bug fixes and extensions

OmniNetworking:
	- Fixed clang-sa and 64-bit issues

OWF:
	- Fixed clang-sa issue

+++ Frameworks/OmniAppKit/Tests/UnitTests/OATestCase.m
@@ -32,7 +32,7 @@ RCS_ID("$Id$");
 
 @implementation OATestController
 
-- (BOOL)exceptionHandler:(NSExceptionHandler *)sender shouldLogException:(NSException *)exception mask:(unsigned int)aMask;
+- (BOOL)exceptionHandler:(NSExceptionHandler *)sender shouldLogException:(NSException *)exception mask:(NSUInteger)aMask;
 {
     // if we are in a shouldRaise or shouldNotRaise, we don't want to get spammed by backtraces.  if we are supposed to raise, that'll be checked and it's "valid".  If we aren't supposed to raise, otest will catch this.
     return NO;
