+++ Frameworks/OmniFoundation/DataStructures.subproj/OFRandom.h
@@ -12,16 +12,16 @@
 // Some platforms don't provide random number generation and of those that do, there are many different variants.  We provide a common random number generator rather than have to deal with each platform independently.  Additionally, we allow the user to maintain several random number generators.
 
 typedef struct {
-    unsigned long y; // current value any number between zero and M-1
+    uint32_t y; // current value any number between zero and M-1
 } OFRandomState;
 
 extern float OFRandomMax;
 
 extern unsigned int OFRandomGenerateRandomSeed(void);	// returns a random number (generated from /dev/urandom if possible, otherwise generated via clock information) for use as a seed value for OFRandomSeed().
-extern void OFRandomSeed(OFRandomState *state, unsigned long y);
-extern unsigned long OFRandomNextState(OFRandomState *state);
+extern void OFRandomSeed(OFRandomState *state, uint32_t y);
+extern uint32_t OFRandomNextState(OFRandomState *state);
 
-extern unsigned long OFRandomNext(void);	// returns a random number generated using a default, shared random state.
+extern uint32_t OFRandomNext(void);	// returns a random number generated using a default, shared random state.
 extern float OFRandomGaussState(OFRandomState *state);
 
 static inline float OFRandomFloat(OFRandomState *state)
